+++ R/install.R
@@ -115,7 +115,7 @@ install_keras <- function(method = c("auto", "virtualenv", "conda"),
     have_conda <- !is.null(tryCatch(conda_binary(conda), error = function(e) NULL))
     if (!have_conda) {
       stop("Keras installation failed (no conda binary found)\n\n",
-           "Install Anaconda for Python 3.x (https://www.continuum.io/downloads#windows)\n",
+           "Install Anaconda for Python 3.x (https://www.anaconda.com/download/#windows)\n",
            "before installing Keras.",
            call. = FALSE)
     }